// src/app/modules/academicSemester/academic.Sem.Model.ts
import httpStatus from 'http-status';
import { model, Schema } from 'mongoose';
import AppError from '../../errors/appError';
import { TAcademicSemester } from './academic.Sem.Interface';
import { Months, semesterCode, semesterName } from './academicSem.constant';

const academicSemesterSchema = new Schema<TAcademicSemester>(
  {
    name: {
      type: String,
      enum: semesterName,
      required: true,
    },
    year: {
      type: String,
      required: true,
    },
    code: {
      type: String,
      enum: semesterCode,
      required: true,
    },
    startMonth: {
      type: String,
      enum: Months,
      required: true,
    },
    endMonth: {
      type: String,
      enum: Months,
      required: true,
    },
  },
  {
    timestamps: true,
  },
);



// check if the semester exists or not
academicSemesterSchema.pre('save', async function (next) {
  const isSemesterExist = await AcademicSemester.findOne({
    name: this.name,
    year: this.year,
  });

  if (isSemesterExist) {
    throw new AppError(httpStatus.NOT_FOUND,'Semester already exists');
  }

  next();
  
});




export const AcademicSemester = model<TAcademicSemester>(
  'AcademicSemester',
  academicSemesterSchema,
);
